+++ packages/ui/src/core/Backdrop/index.ts
@@ -1,0 +1,2 @@
+export { default } from './backdrop.component';
+export type { BackdropProps } from './backdrop.props';