+++ packages/ui/src/core/calendar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './calendar.component';
+export type { CalendarProps } from './calendar.props';