+++ packages/ui/src/core/date-picker/index.ts
@@ -1,0 +1,2 @@
+export { default } from './date-picker.component';
+export type { DatePickerProps } from './date-picker.props';