+++ packages/ui/src/core/drawer/drawer-item.component.tsx
@@ -1,0 +1,28 @@
+import React, { useContext } from 'react';
+import PT from 'prop-types';
+
+import { intrinsicComponent } from '../../utils/functions';
+import type { DrawerItemProps } from './drawer.props';
+import DrawerContext from './drawer.context';
+import Styled from './drawer.styles';
+
+const DrawerItem = intrinsicComponent<DrawerItemProps, HTMLLIElement>(
+  ({ children, selected, ...rest }, ref): JSX.Element => {
+    const { isCollapsed } = useContext(DrawerContext);
+    return (
+      <Styled.Item selected={selected} isCollapsed={isCollapsed} {...rest} ref={ref}>
+        {children}
+      </Styled.Item>
+    );
+  }
+);
+
+DrawerItem.defaultProps = {
+  selected: false,
+};
+
+DrawerItem.propTypes = {
+  selected: PT.bool,
+};
+
+export default DrawerItem;