+++ packages/ui/src/core/index.ts
@@ -57,4 +57,8 @@ export { default as Textarea } from './textarea';
 export { default as Tooltip } from './tooltip';
 export { default as UploadInput } from './upload-input';
 export { default as Skeleton } from './skeleton';
+export { default as Notification } from './notification';
+export { default as Calendar } from './calendar';
+export { default as DatePicker } from './date-picker';
 export { default as CrossButton } from './cross-button';
+export { default as Divider } from './divider';